+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.8.2 - Unreleased
 
+### Added
+* Added option to mask nodata regions when using imread and gdal backend.
+
 
 ## Version 0.8.1 - Released 2022-03-04